[DEV] libulog.h: add more errno report macros
Add macros to report errno errors at different levels.

It makes possible to report errors according to their
importance, without to repeat the format of ULOG_ERRNO.

Example:

    err = function(arg);
    switch (err)
    {
    case -EFAULT:
        level = ULOG_CRI;
        break;
    case -EINVAL:
        level = ULOG_ERR;
        break;
    default:
        level = ULOG_DBG;
        break;
    }
    ULOG_PRI_ERRNO(err, level, "failed function(%d)", arg);

    err = function(arg);
    if (err == -EAGAIN)
        ULOGD_ERRNO(err, "temporary failed function(%d)", arg);
    else if (err)
        ULOGE_ERRNO(err, "fatal failed function(%d)", arg);

The order of parameter is reversed compared to existing
ULOG_ERRNO(), first because it allows for the format to
to be directly followed by its arguments, second to match
the ULOG_THROTTLE and ULOG_CHANGE scheme.

Change-Id: I21e22b6e89abb933959b636365bd6e61d8110292
Reviewed-on: https://gerrit-doc.parrot.biz/c/pulsar/ulog/+/149508
Tested-by: Jenkins_dragon <jenkins.dragon@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>